2nd attempt to manage floats

+++ db/cats_db.py
@@ -1,6 +1,6 @@
 from tokenize import Floatnumber
 from sqlalchemy import Column, ForeignKey
-from sqlalchemy import Integer, String
+from sqlalchemy import Integer, Float, String
 from sqlalchemy.sql.sqltypes import Boolean
 from sqlalchemy import Date
 from db.db_connection import Base, engine
@@ -10,8 +10,8 @@ class CatsInDb(Base):
     category    = Column(String, primary_key=True)
     type        = Column(String)
     username    = Column(String, ForeignKey("users.username"))
-    budget      = Column(Integer)
-    value       = Column(Integer)
+    budget      = Column(Float)
+    value       = Column(Float)
     recurrency  = Column(Boolean)
     day         = Column(Date, nullable=True )
