+++ models/goal_models.py
@@ -4,13 +4,13 @@ from datetime import date
 class GoalIn (BaseModel):
     name        : str
     username    : str
-    final_value : int
+    final_value : float
     final_date  : date
 
 class GoalUp (BaseModel):
     name        : str
     username    : str
-    value       : int
+    value       : float
 
 class GoalDel (BaseModel):
     username    : str
@@ -18,8 +18,8 @@ class GoalDel (BaseModel):
 
 class GoalOut (BaseModel):
     name        : str
-    current_val : int
-    final_value : int
+    current_val : float
+    final_value : float
     final_date  : date
 
     class Config:
